flask-nextjs-chat/backend/requirements.txt
bidict==0.23.1
blinker==1.8.2
certifi==2024.7.4
chardet==3.0.4
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
eventlet==0.36.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
fpdf==1.7.2
googletrans==3.1.0a0
greenlet==3.0.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.7.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
PyPDF2==3.0.1
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
rfc3986==1.5.0
simple-websocket==1.0.0
sniffio==1.3.1
Werkzeug==3.0.3
wsproto==1.2.0
